feat: add vector3d library
