+++ lib/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(deltaTime)
 add_subdirectory(window)
+add_subdirectory(vector3d)
