+++ src/physics/rigidBody/rigidBody.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "vector3d.hpp"
 
 #define PI 3.14f
 
@@ -12,27 +13,6 @@ public:
 
 	virtual float volume() = 0;
 	virtual float surfaceArea() = 0;
-};
-
-class Vector3d {
-public:
-	Vector3d(): x(0), y(0), z(0) {};
-
-	Vector3d(float x, float y, float z);
-	~Vector3d(){};
-
-	float x;
-	float y;
-	float z;
-
-	Vector3d crossProduct(const Vector3d& vec);
-	float dotProduct(const Vector3d& vec);
-	float magnitude();
-
-	// TODO:: https://people.math.harvard.edu/~jjchen/math21a/handouts/vector-ops.html
-	// - [ ] Vector addition
-	// - [ ] Scalar multiplication
-	// - [ ] Transformations?
 };
 
 class RigidBody {
